+++ src/Bus/CommandHandlerInterface.php
@@ -19,6 +19,8 @@ declare(strict_types=1);
 
 namespace CloudCreativity\Modules\Bus;
 
+use CloudCreativity\Modules\Toolkit\Messages\CommandInterface;
+use CloudCreativity\Modules\Toolkit\Messages\DispatchThroughMiddleware;
 use CloudCreativity\Modules\Toolkit\Result\ResultInterface;
 
 interface CommandHandlerInterface extends DispatchThroughMiddleware