+++ src/Bus/Middleware/LogMessageDispatch.php
@@ -20,9 +20,9 @@ declare(strict_types=1);
 namespace CloudCreativity\Modules\Bus\Middleware;
 
 use Closure;
-use CloudCreativity\Modules\Bus\MessageInterface;
 use CloudCreativity\Modules\Infrastructure\Log\ObjectContext;
 use CloudCreativity\Modules\Infrastructure\Log\ResultContext;
+use CloudCreativity\Modules\Toolkit\Messages\MessageInterface;
 use CloudCreativity\Modules\Toolkit\ModuleBasename;
 use CloudCreativity\Modules\Toolkit\Result\ResultInterface;
 use Psr\Log\LoggerInterface;