+++ src/Bus/Middleware/ValidateCommand.php
@@ -20,8 +20,8 @@ declare(strict_types=1);
 namespace CloudCreativity\Modules\Bus\Middleware;
 
 use Closure;
-use CloudCreativity\Modules\Bus\CommandInterface;
 use CloudCreativity\Modules\Bus\Validation\CommandValidatorInterface;
+use CloudCreativity\Modules\Toolkit\Messages\CommandInterface;
 use CloudCreativity\Modules\Toolkit\Result\Result;
 use CloudCreativity\Modules\Toolkit\Result\ResultInterface;