+++ src/Bus/Middleware/ValidateQuery.php
@@ -20,8 +20,8 @@ declare(strict_types=1);
 namespace CloudCreativity\Modules\Bus\Middleware;
 
 use Closure;
-use CloudCreativity\Modules\Bus\QueryInterface;
 use CloudCreativity\Modules\Bus\Validation\QueryValidatorInterface;
+use CloudCreativity\Modules\Toolkit\Messages\QueryInterface;
 use CloudCreativity\Modules\Toolkit\Result\Result;
 use CloudCreativity\Modules\Toolkit\Result\ResultInterface;