+++ src/EventBus/IntegrationEventHandler.php
@@ -1,0 +1,67 @@
+<?php
+/*
+ * Copyright 2024 Cloud Creativity Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+declare(strict_types=1);
+
+namespace CloudCreativity\Modules\EventBus;
+
+use Closure;
+use CloudCreativity\Modules\Toolkit\Messages\DispatchThroughMiddleware;
+use CloudCreativity\Modules\Toolkit\Messages\IntegrationEventInterface;
+
+final class IntegrationEventHandler implements IntegrationEventHandlerInterface
+{
+    /**
+     * IntegrationEventHandler constructor.
+     *
+     * @param object $handler
+     */
+    public function __construct(private readonly object $handler)
+    {
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function __invoke(IntegrationEventInterface $event): void
+    {
+        if ($this->handler instanceof Closure) {
+            ($this->handler)($event);
+            return;
+        }
+
+        assert(method_exists($this->handler, 'handle'), sprintf(
+            'Cannot handle "%s" - handler "%s" does not have a handle method.',
+            $event::class,
+            $this->handler::class,
+        ));
+
+        $this->handler->handle($event);
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function middleware(): array
+    {
+        if ($this->handler instanceof DispatchThroughMiddleware) {
+            return $this->handler->middleware();
+        }
+
+        return [];
+    }
+}