+++ CHANGELOG.md
@@ -5,6 +5,41 @@ All notable changes to this project will be documented in this file. This projec
 
 ## Unreleased
 
+## [1.0.0-rc.1] - 2024-02-23
+
+### Added
+
+- New event bus notifier implementation that was previously missing. This completes the event bus implementation.
+- New message interfaces (command, query, integration event) added to the toolkit.
+- New loggable context provider interface added to the toolkit.
+- Module basename now supports namespaces where an application only has a single bounded context.
+
+### Changed
+
+- **BREAKING** Moved the following interfaces to the `Toolkit\Messages` namespace:
+    - `MessageInterface`
+    - `IntegrationEventInterface`
+- **BREAKING** Interfaces that type-hinted `Bus\CommandInterface`, `Bus\QueryInterface` or `Bus\MessageInterface` now
+  type-hint the new interfaces in the `Toolkit\Messages` namespace.
+- **BREAKING** Moved the `EventBus` implementation from `Infrastructure\EventBus` to `EventBus`. In Deptrac, this
+  namespace is now part of the _Application Bus_ layer. Renamed the publisher handler and publisher handler containers
+  to integration event handler and container - so that they can be used for both the publisher and notifier
+  implementations.
+- **BREAKING** Removed the `EventBus\PublishThroughMiddleware` interface. Use the
+  `Toolkit\Messages\DispatchThroughMiddleware` interface instead.
+
+### Removed
+
+- **BREAKING** removed the `deptrac-layers.yaml` file, in favour of applications including the classes in their own
+   Deptrac configuration.
+
+### Deprecated
+
+- The `Bus\CommandInterface`, `Bus\QueryInterface` and `Bus\DispatchThroughMiddleware` interfaces have been deprecated
+  in favour of the new interfaces in the `Toolkit\Messages` namespace.
+- The `Infrastructure\Log\ContextProviderInterface` is deprecated in favour of the new
+  `Toolkit\Loggable\ContextProviderInterface` interface.
+
 ## [0.6.1] - 2024-02-09
 
 ### Fixed